it runs but without anything on the homeform

+++ src/homeform.cpp
@@ -530,44 +530,51 @@ homeform::homeform(QQmlApplicationEngine *engine, bluetooth *bl) {
     connect(backupTimer, &QTimer::timeout, this, &homeform::backup);
     backupTimer->start(1min);
 
-    QObject *rootObject = engine->rootObjects().constFirst();
-    QObject *home = rootObject->findChild<QObject *>(QStringLiteral("home"));
-    QObject *stack = rootObject;
-    engine->rootContext()->setContextProperty("pathController", &pathController);
-    QObject::connect(home, SIGNAL(start_clicked()), this, SLOT(Start()));
-    QObject::connect(home, SIGNAL(stop_clicked()), this, SLOT(Stop()));
-    QObject::connect(stack, SIGNAL(trainprogram_open_clicked(QUrl)), this, SLOT(trainprogram_open_clicked(QUrl)));
-    QObject::connect(stack, SIGNAL(trainprogram_open_other_folder(QUrl)), this, SLOT(trainprogram_open_other_folder(QUrl)));
-    QObject::connect(stack, SIGNAL(gpx_open_other_folder(QUrl)), this, SLOT(gpx_open_other_folder(QUrl)));
-    QObject::connect(stack, SIGNAL(profile_open_clicked(QUrl)), this, SLOT(profile_open_clicked(QUrl)));
-    QObject::connect(stack, SIGNAL(trainprogram_preview(QUrl)), this, SLOT(trainprogram_preview(QUrl)));
-    QObject::connect(stack, SIGNAL(gpxpreview_open_clicked(QUrl)), this, SLOT(gpxpreview_open_clicked(QUrl)));
-    QObject::connect(stack, SIGNAL(trainprogram_zwo_loaded(QString)), this, SLOT(trainprogram_zwo_loaded(QString)));
-    QObject::connect(stack, SIGNAL(gpx_open_clicked(QUrl)), this, SLOT(gpx_open_clicked(QUrl)));
-    QObject::connect(stack, SIGNAL(gpx_save_clicked()), this, SLOT(gpx_save_clicked()));
-    QObject::connect(stack, SIGNAL(fit_save_clicked()), this, SLOT(fit_save_clicked()));
-    QObject::connect(stack, SIGNAL(strava_connect_clicked()), this, SLOT(strava_connect_clicked()));
-    QObject::connect(stack, SIGNAL(refresh_bluetooth_devices_clicked()), this,
-                     SLOT(refresh_bluetooth_devices_clicked()));
-    QObject::connect(home, SIGNAL(lap_clicked()), this, SLOT(Lap()));
-    QObject::connect(home, SIGNAL(peloton_start_workout()), this, SLOT(peloton_start_workout()));
-    QObject::connect(home, SIGNAL(peloton_abort_workout()), this, SLOT(peloton_abort_workout()));
-    QObject::connect(stack, SIGNAL(loadSettings(QUrl)), this, SLOT(loadSettings(QUrl)));
-    QObject::connect(stack, SIGNAL(saveSettings(QUrl)), this, SLOT(saveSettings(QUrl)));
-    QObject::connect(stack, SIGNAL(deleteSettings(QUrl)), this, SLOT(deleteSettings(QUrl)));
-    QObject::connect(stack, SIGNAL(restoreSettings()), this, SLOT(restoreSettings()));
-    QObject::connect(stack, SIGNAL(saveProfile(QString)), this, SLOT(saveProfile(QString)));
-    QObject::connect(stack, SIGNAL(restart()), this, SLOT(restart()));
+    QObject *stack = nullptr;
 
-    QObject::connect(stack, SIGNAL(volumeUp()), this, SLOT(volumeUp()));
-    QObject::connect(stack, SIGNAL(volumeDown()), this, SLOT(volumeDown()));
-    QObject::connect(stack, SIGNAL(keyMediaPrevious()), this, SLOT(keyMediaPrevious()));
-    QObject::connect(stack, SIGNAL(keyMediaNext()), this, SLOT(keyMediaNext()));
-    QObject::connect(stack, SIGNAL(floatingOpen()), this, SLOT(floatingOpen()));
-    QObject::connect(stack, SIGNAL(openFloatingWindowBrowser()), this, SLOT(openFloatingWindowBrowser()));
-    QObject::connect(stack, SIGNAL(strava_upload_file_prepare()), this, SLOT(strava_upload_file_prepare()));
+    if(engine->rootObjects().count() > 0) {
+        QObject *rootObject = engine->rootObjects().constFirst();
+        QObject *home = rootObject->findChild<QObject *>(QStringLiteral("home"));
+        QObject *stack = rootObject;
 
-    qDebug() << "homeform constructor events linked";
+        engine->rootContext()->setContextProperty("pathController", &pathController);
+        QObject::connect(home, SIGNAL(start_clicked()), this, SLOT(Start()));
+        QObject::connect(home, SIGNAL(stop_clicked()), this, SLOT(Stop()));
+        QObject::connect(stack, SIGNAL(trainprogram_open_clicked(QUrl)), this, SLOT(trainprogram_open_clicked(QUrl)));
+        QObject::connect(stack, SIGNAL(trainprogram_open_other_folder(QUrl)), this, SLOT(trainprogram_open_other_folder(QUrl)));
+        QObject::connect(stack, SIGNAL(gpx_open_other_folder(QUrl)), this, SLOT(gpx_open_other_folder(QUrl)));
+        QObject::connect(stack, SIGNAL(profile_open_clicked(QUrl)), this, SLOT(profile_open_clicked(QUrl)));
+        QObject::connect(stack, SIGNAL(trainprogram_preview(QUrl)), this, SLOT(trainprogram_preview(QUrl)));
+        QObject::connect(stack, SIGNAL(gpxpreview_open_clicked(QUrl)), this, SLOT(gpxpreview_open_clicked(QUrl)));
+        QObject::connect(stack, SIGNAL(trainprogram_zwo_loaded(QString)), this, SLOT(trainprogram_zwo_loaded(QString)));
+        QObject::connect(stack, SIGNAL(gpx_open_clicked(QUrl)), this, SLOT(gpx_open_clicked(QUrl)));
+        QObject::connect(stack, SIGNAL(gpx_save_clicked()), this, SLOT(gpx_save_clicked()));
+        QObject::connect(stack, SIGNAL(fit_save_clicked()), this, SLOT(fit_save_clicked()));
+        QObject::connect(stack, SIGNAL(strava_connect_clicked()), this, SLOT(strava_connect_clicked()));
+        QObject::connect(stack, SIGNAL(refresh_bluetooth_devices_clicked()), this,
+                         SLOT(refresh_bluetooth_devices_clicked()));
+        QObject::connect(home, SIGNAL(lap_clicked()), this, SLOT(Lap()));
+        QObject::connect(home, SIGNAL(peloton_start_workout()), this, SLOT(peloton_start_workout()));
+        QObject::connect(home, SIGNAL(peloton_abort_workout()), this, SLOT(peloton_abort_workout()));
+        QObject::connect(stack, SIGNAL(loadSettings(QUrl)), this, SLOT(loadSettings(QUrl)));
+        QObject::connect(stack, SIGNAL(saveSettings(QUrl)), this, SLOT(saveSettings(QUrl)));
+        QObject::connect(stack, SIGNAL(deleteSettings(QUrl)), this, SLOT(deleteSettings(QUrl)));
+        QObject::connect(stack, SIGNAL(restoreSettings()), this, SLOT(restoreSettings()));
+        QObject::connect(stack, SIGNAL(saveProfile(QString)), this, SLOT(saveProfile(QString)));
+        QObject::connect(stack, SIGNAL(restart()), this, SLOT(restart()));
+
+        QObject::connect(stack, SIGNAL(volumeUp()), this, SLOT(volumeUp()));
+        QObject::connect(stack, SIGNAL(volumeDown()), this, SLOT(volumeDown()));
+        QObject::connect(stack, SIGNAL(keyMediaPrevious()), this, SLOT(keyMediaPrevious()));
+        QObject::connect(stack, SIGNAL(keyMediaNext()), this, SLOT(keyMediaNext()));
+        QObject::connect(stack, SIGNAL(floatingOpen()), this, SLOT(floatingOpen()));
+        QObject::connect(stack, SIGNAL(openFloatingWindowBrowser()), this, SLOT(openFloatingWindowBrowser()));
+        QObject::connect(stack, SIGNAL(strava_upload_file_prepare()), this, SLOT(strava_upload_file_prepare()));
+
+        qDebug() << "homeform constructor events linked";
+    } else {
+        qDebug() << "error on QML engine UI";
+    }
 
 #if !defined(Q_OS_ANDROID) && !defined(Q_OS_IOS)
     QObject::connect(engine, &QQmlApplicationEngine::quit, &QGuiApplication::quit);
@@ -591,7 +598,8 @@ homeform::homeform(QQmlApplicationEngine *engine, bluetooth *bl) {
     connect(pelotonHandler, &peloton::pzpLoginState, this, &homeform::pzpLoginState);
     connect(pelotonHandler, &peloton::pelotonAuthUrlChanged, this, &homeform::pelotonAuthUrlChanged);
     connect(pelotonHandler, &peloton::pelotonWebVisibleChanged, this, &homeform::pelotonWebVisibleChanged);
-    connect(stack, SIGNAL(peloton_connect_clicked()), pelotonHandler, SLOT(peloton_connect_clicked()));
+    if(stack)
+        connect(stack, SIGNAL(peloton_connect_clicked()), pelotonHandler, SLOT(peloton_connect_clicked()));
 
     // copying bundles zwo files in the right path if necessary
     QDirIterator itZwo(":/zwo/");
